try compatibility with pre 1.15 as noop
