+++ proxy_https_legacy.go
@@ -1,0 +1,15 @@
+// +build !go1.15
+
+package websocket
+
+import (
+	"context"
+	"net/url"
+)
+
+func registerDialerHttps() {
+}
+
+func modifyProxyDialer(ctx context.Context, d *Dialer, proxyURL *url.URL, proxyDialer *netDialerFunc) {
+	return nil
+}
